:recycle: refactor: discount, update 로직별 파일 분리
