--- src/basic/utils/index.js
@@ -1,1 +1,0 @@
-export * from "./discountUtil";
